Move rich related terminal code into own module
